Move StateUI component
